Store various developments to IO speed-up - not yet complete
refs #142


git-svn-id: https://svn.isis.rl.ac.uk/Horace/branches/HoraceV3.1.1@935 1cb955da-4463-431d-88dd-2af2f5550de3

Former-commit-id: 704f7a8277f5374ca47494930c33694527a25ce7

+++ @sqw/private/get_sqw_data_signal_sparse.m
@@ -248,9 +248,9 @@ try
             tmp = fread(fid, npix_read, ['*',fmt.pix]);
             if make_full_fmt
                 if datastruct
-                    data.pix = pix_sparse_to_full(tmp,pix_nz,ind_beg,max(info.ne),info.ndet);
+                    data.pix = pix_sparse_to_full(double(tmp),pix_nz,ind_beg,max(info.ne),info.ndet);
                 else
-                    data = pix_sparse_to_full(tmp,pix_nz,ind_beg,max(info.ne),info.ndet);
+                    data = pix_sparse_to_full(double(tmp),pix_nz,ind_beg,max(info.ne),info.ndet);
                 end
             else
                 if datastruct
